requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.6
APScheduler==3.10.4
python-dateutil==2.9.0.post0
